requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.9.1
attrs==25.3.0
bce-python-sdk==0.9.56
beautifulsoup4==4.13.5
CacheControl==0.14.3
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.1
colorlog==6.10.1
cryptography==45.0.7
Deprecated==1.2.18
deprecation==2.1.0
Django==5.2.6
easyocr==1.7.2
et_xmlfile==2.0.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
fuzzywuzzy==0.18.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.10
imageio==2.37.0
imagesize==1.4.1
img2pdf==0.6.1
Jinja2==3.1.6
lazy_loader==0.4
Levenshtein==0.27.1
lxml==6.0.1
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
modelscope==1.33.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.4
networkx==3.5
ninja==1.13.0
numpy==2.2.6
ocrmypdf==16.10.4
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.2
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.0
pikepdf==9.10.2
pillow==11.3.0
playwright==1.55.0
pluggy==1.6.0
prettytable==3.17.0
propcache==0.3.2
protobuf==6.33.2
psutil==7.1.3
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
PyMuPDF==1.26.4
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyobjc-framework-Quartz==12.1
pypdfium2==5.2.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.0
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
rubicon-objc==0.5.2
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
sympy==1.14.0
tifffile==2025.9.9
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
wrapt==1.17.3
yarl==1.20.1
